# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AI agent boilerplate — public API",
    "description": "The browser-facing API (auth, searches, recurring searches). Contract pinned by fixtures + zod (ADR-049).",
    "license": {
      "name": "MIT"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/auth/refresh": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Rotates the refresh cookie and returns a fresh access token (ADR-008).",
        "operationId": "refresh",
        "responses": {
          "200": {
            "description": "Rotated refresh cookie, new access token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccessTokenResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid refresh cookie",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AccessTokenResponse": {
        "type": "object",
        "description": "`{ \"access_token\": \"…\" }` — the short-lived bearer; the refresh token is set\nas an HttpOnly cookie (ADR-008), not in the body.",
        "required": [
          "access_token"
        ],
        "properties": {
          "access_token": {
            "type": "string"
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "description": "`{ \"error\": \"…\" }` — the uniform error body (ADR-018).",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
}